requirements.txt
Faker==19.3.0
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
